Support str and bytes
